Refactor match card to composables file for reusability
